fix: [FFM-5118]: Refactor using PR comments in initial review (#12117)
