+++ src/modules/75-cf/pages/home/trialPage/CFTrialPage.module.scss
@@ -28,36 +28,36 @@
 
 .mainArticle {
   grid-template-columns: 1fr 1px 1fr;
-  gap: 3rem;
+  gap: var(--spacing-xlarge);
   width: 800px;
   padding: var(--spacing-xxxlarge);
   background-color: var(--grey-0);
-  border-radius: 0.5rem;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1), 0 5px 20px rgba(0, 0, 0, 0.1);
+  border-radius: var(--spacing-small);
+  box-shadow: var(--elevation-3);
   display: grid;
 }
 
 .dontCodeArticle {
   grid-template-columns: 1fr 2fr;
-  gap: 3rem;
+  gap: var(--spacing-xlarge);
   width: 800px;
   padding: var(--spacing-xxxlarge);
   background-color: var(--grey-0);
-  border-radius: 0.5rem;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1), 0 5px 20px rgba(0, 0, 0, 0.1);
+  border-radius: var(--spacing-small);
+  box-shadow: var(--elevation-3);
   display: grid;
 }
 
 .forDevs {
   display: flex;
   flex-direction: column;
-  gap: 1rem;
+  gap: var(--spacing-medium);
 }
 
 .dontCode {
   display: flex;
   flex-direction: column;
-  gap: 0.5rem;
+  gap: var(--spacing-small);
   width: 450px;
 }
 
@@ -67,33 +67,32 @@
 }
 
 .divider {
-  background-color: var(--grey-200) !important;
+  background-color: var(--grey-200);
 }
 
 .startFreePlanBtn {
   grid-column: 1/4;
   justify-self: center;
-  border-radius: 0.2rem;
 }
 
 .list {
   list-style: none;
   display: flex;
   flex-direction: column;
-  gap: 1rem;
-  padding: 0 !important;
+  gap: var(--spacing-medium);
+  padding: 0;
   align-items: flex-start;
 
   li {
     display: grid;
     grid-template-columns: min-content 1fr;
     align-items: center;
-    gap: 1rem;
-    width: 280px;
+    gap: var(--spacing-medium);
   }
 
   li::before {
-    content: url('./star.svg');
-    font-size: 2rem;
+    content: url('../star.svg');
+    height: 28px;
+    width: 31px;
   }
 }
